Moved UUID extension methods to its own extension package
