+++ src/main/kotlin/ch/hevs/cloudio/cloud/extension/UUID.kt
@@ -1,4 +1,4 @@
-package ch.hevs.cloudio.cloud.utils
+package ch.hevs.cloudio.cloud.extension
 
 import java.math.BigInteger
 import java.util.*
